LPD-93941 Remove CMP dependency from CMS site initializer tests
The site-cms-site-initializer-test module depended on
site-cmp-site-initializer-test-util through CMPTestUtil, coupling the CMS
test bundle to the CMP product. The two affected tests use CMP differently,
so they are handled differently.

ViewSpaceMembersSummarySectionDisplayContextTest only needs an object entry
in a TYPE_PROJECT depot to exercise the display context, since the production
code branches solely on the depot type. It now builds a TYPE_PROJECT depot
with a CMS basic web content entry instead of a CMP project, removing the CMP
dependency while preserving the same code path.

ObjectEntryModelListenerTest mixed CMS resource-permission scenarios with CMP
audit-routing scenarios. The audit routing genuinely requires the CMP task
object definition, so those scenarios move to a new ObjectEntryAuditModelListenerTest
in site-cmp-site-initializer-test, where the CMP concepts are owned. The CMS
test keeps only the resource-permission coverage and no longer references CMP.

With both tests cleaned up, the CMP test-util dependency is dropped from the
CMS test module.
