+++ modules/apps/site/site-cms-site-initializer-test/src/testIntegration/java/com/liferay/site/cms/site/initializer/internal/display/context/test/ViewSpaceMembersSummarySectionDisplayContextTest.java
@@ -11,13 +11,11 @@ import com.liferay.depot.model.DepotEntry;
 import com.liferay.depot.service.DepotEntryLocalService;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.info.constants.InfoDisplayWebKeys;
-import com.liferay.object.model.ObjectEntry;
 import com.liferay.object.rest.test.util.ObjectEntryTestUtil;
 import com.liferay.object.service.ObjectDefinitionLocalService;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.test.AssertUtils;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -31,14 +29,12 @@ import com.liferay.portal.test.rule.FeatureFlags;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
-import com.liferay.site.cmp.site.initializer.test.util.CMPTestUtil;
 
 import java.io.Serializable;
 
 import java.util.List;
 import java.util.Map;
 
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -62,15 +58,6 @@ public class ViewSpaceMembersSummarySectionDisplayContextTest
 		new AggregateTestRule(
 			new LiferayIntegrationTestRule(),
 			PermissionCheckerMethodTestRule.INSTANCE);
-
-	@Before
-	@Override
-	public void setUp() throws Exception {
-		super.setUp();
-
-		CMPTestUtil.getOrAddGroup(
-			ViewSpaceMembersSummarySectionDisplayContextTest.class);
-	}
 
 	@Test
 	public void testGetHeaderProps() throws Exception {
@@ -96,10 +83,25 @@ public class ViewSpaceMembersSummarySectionDisplayContextTest
 
 		_assertHeaderProps(null, depotEntry.getGroup());
 
-		ObjectEntry objectEntry = CMPTestUtil.addProjectObjectEntry();
+		DepotEntry projectDepotEntry = _depotEntryLocalService.addDepotEntry(
+			RandomTestUtil.randomLocaleStringMap(),
+			RandomTestUtil.randomLocaleStringMap(), DepotConstants.TYPE_PROJECT,
+			ServiceContextTestUtil.getServiceContext());
 
 		mockHttpServletRequest.setAttribute(
-			InfoDisplayWebKeys.INFO_ITEM, objectEntry);
+			InfoDisplayWebKeys.INFO_ITEM,
+			ObjectEntryTestUtil.addObjectEntry(
+				projectDepotEntry.getGroupId(),
+				_objectDefinitionLocalService.
+					fetchObjectDefinitionByExternalReferenceCode(
+						"L_CMS_BASIC_WEB_CONTENT",
+						TestPropsValues.getCompanyId()),
+				HashMapBuilder.<String, Serializable>put(
+					"title_i18n",
+					HashMapBuilder.put(
+						"en_US", RandomTestUtil.randomString()
+					).build()
+				).build()));
 
 		_assertHeaderProps(
 			HashMapBuilder.<String, Object>put(
@@ -113,7 +115,7 @@ public class ViewSpaceMembersSummarySectionDisplayContextTest
 						StringUtil.merge(depotEntryGroupIds) + ")";
 				}
 			).build(),
-			_groupLocalService.fetchGroup(objectEntry.getGroupId()));
+			projectDepotEntry.getGroup());
 	}
 
 	private void _assertHeaderProps(
@@ -173,9 +175,6 @@ public class ViewSpaceMembersSummarySectionDisplayContextTest
 		filter = "component.name=com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewSpaceMembersSummaryJSPSectionFragmentRenderer"
 	)
 	private FragmentRenderer _fragmentRenderer;
-
-	@Inject
-	private GroupLocalService _groupLocalService;
 
 	@Inject
 	private ObjectDefinitionLocalService _objectDefinitionLocalService;
